More web friendly and changed nav in mobile site
Made the photos smaller size and changed codex on them to be opened
easier on websites. Changed the KenKou logo change in size and padding
depending a width of page using media query. Changed the navigation on
the mobile site to the right hand side and flipped colours.
Also added mobile redirect on main site homepage (untested so far)